feat: open governance referendum via link
